refactor schema: change data types for time, duration, order_index, and repetitions to smaller types

+++ src/main/java/source/code/dto/request/workoutSet/WorkoutSetCreateDto.java
@@ -12,10 +12,10 @@ import lombok.Setter;
 @NoArgsConstructor
 public class WorkoutSetCreateDto {
     @NotNull
-    private Integer orderIndex;
+    private Short orderIndex;
 
     @NotNull
-    private Integer restSeconds;
+    private Short restSeconds;
 
     @NotNull
     private Integer workoutId;
